Add ReportDbContext  + migration

+++ ReportsManagements/Models/TrainerReport.cs
@@ -8,7 +8,7 @@ namespace ReportsManagements.Models
         public int TrainerReportId { get; set; }
         public int TotalSessions { get; set; }
         public int TotalStudents { get; set; }
-        public int AttendanceRate { get; set; }
+        public decimal AttendanceRate { get; set; }
         public int TrainerId { get; set; }
         public int CourseId { get; set; }
     }
